requirements.txt
click==8.1.3
et-xmlfile==1.1.0
Flask==2.2.3
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.2
openpyxl==3.1.1
pandas~=1.3.5
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
Werkzeug==2.2.3
xlrd==2.0.1
zipp==3.15.0
certifi==2021.10.8
click==8.0.3
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.21.2
pandas==1.3.5
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
Werkzeug==2.0.2

dash~=2.8.1
plotly~=5.13.1
APScheduler~=3.10.0
Pillow~=9.4.0
pyserial~=3.5